Changes for avatar select menu
