+++ packages/client-core/src/common/components/AvatarPreview/index.module.scss
@@ -1,0 +1,7 @@
+.preview {
+  color: var(--textColor);
+  border: 1px solid var(--buttonOutlined);
+  border-radius: 8px;
+  box-shadow: 0 0 10px var(--buttonOutlined);
+  position: relative;
+}
